requirements.txt
fastapi~=0.115.12
uvicorn
numpy~=1.26.4
opencv-python-headless
pillow~=11.0.0
keras~=3.6.0
tensorflow~=2.18.0

opencv-python~=4.10.0.84
pydantic~=2.11.4
matplotlib~=3.9.2
requests~=2.32.3
pandas~=2.2.3